# Faster cde rejection