@Pr0methean 2024-04-30
Because this repo is no longer maintained, I've replaced this PR with https://github.com/zip-rs/zip2/pull/47.